Refactor and Docker File
